+++ internal/pidfile/process_windows.go
@@ -2,7 +2,10 @@
 
 package pidfile
 
-import "os"
+import (
+	"os"
+	"time"
+)
 
 func IsRunning(pid int) bool {
 	p, err := os.FindProcess(pid)
@@ -16,4 +19,18 @@ func IsRunning(pid int) bool {
 		return true
 	}
 	return false
+}
+
+// GracefulKill on Windows falls back to Kill since SIGTERM is not supported.
+func GracefulKill(pid int, _ time.Duration) error {
+	if !IsRunning(pid) {
+		return nil
+	}
+	p, err := os.FindProcess(pid)
+	if err != nil {
+		return nil
+	}
+	_ = p.Kill()
+	_, _ = p.Wait()
+	return nil
 }